requirements.txt
# Core MCP Dependencies
mcp==1.0.0
httpx>=0.27.0
python-dotenv==1.0.1
pydantic==2.10.3
typing-extensions>=4.7.0

# LangChain + Agent Framework (Stable, Compatible Set)
langchain==0.2.17
langchain-core==0.2.43
langgraph==0.2.35
langchain-groq==0.1.10

# Utilities
requests>=2.31.0
